wangzhan/requirements.txt
# ============================================
# 柑橘检测多智能体网站依赖库
# 推荐使用环境：CUDA 12.8 + Python 3.11+
# ============================================

# Web 框架
flask>=3.0.0
flask-cors>=4.0.0

# PyTorch (GPU 加速版本 - CUDA 12.8)
# 安装命令：pip install torch torchvision --index-url https://download.pytorch.org/whl/cu128
torch>=2.10.0
torchvision>=0.19.0

# 深度学习模型库
timm>=0.9.0

# 图像处理
Pillow>=10.1.0
opencv-python>=4.8.1.78

# RAG 和向量搜索
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4

# HTTP 请求
requests>=2.31.0
httpx>=0.25.2

# 工具库
python-dotenv>=1.0.0
numpy>=1.24.0
scikit-learn>=1.3.2

# Transformers（用于 RAG）
transformers>=4.36.0

# PDF 处理
PyPDF2>=3.0.0

# Word 处理
python-docx>=1.0.0

# OCR（图片文字识别）
# paddlepaddle>=2.5.0  # GPU 版本
# paddlepaddle-cpu>=2.5.0  # CPU 版本
# paddleocr>=2.7.0

# 进度条
tqdm>=4.65.0

# 图像质量评估
scikit-image>=0.21.0

PyPDF2>=3.0.0
python-docx>=1.0.0
paddlepaddle>=2.5.0
paddleocr>=2.7.0
